move test to external tests
